doc_processing_service/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
transformers
Pillow
pdf2image
sentencepiece
timm
# protobuf is often a dependency of sentencepiece or transformers, adding preemptively
protobuf
uvicorn
pytesseract
tesseract
google-cloud
google-api-core
google-cloud-documentai
surya-ocr